imprive buildx script

+++ scripts/buildx.sh
@@ -5,18 +5,21 @@ set -ueo pipefail
 
 SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
-REGISTRY="docker.io/opopops"
-
 BUILDX_ARGS=""
-BUILDX_BUILDER_NAME="opopops"
-BUILDX_BUILDER_PLATFORM="linux/amd64,linux/arm64"
+BUILDX_PRUNE="false"
+BUILDX_BUILDER_NAME="${BUILDX_BUILDER_NAME:-opopops}"
+BUILDX_BUILDER_PLATFORM="${BUILDX_BUILDER_PLATFORM:-linux/amd64,linux/arm64}"
+BUILDX_PRUNE_FILTER="${BUILDX_PRUNE_FILTER:-until=72h}"
+BUILDX_PRUNE_KEEP_STORAGE="${BUILDX_PRUNE_KEEP_STORAGE:-10gb}"
 
 function _exit() {
-  docker buildx prune \
-    --force \
-    --builder="$BUILDX_BUILDER_NAME" \
-    --filter="until=48h" \
-    --keep-storage="2gb"
+  if [[ "$BUILDX_PRUNE" == "true" ]]; then
+    docker buildx prune \
+      --builder="$BUILDX_BUILDER_NAME" \
+      --force \
+      --filter="$BUILDX_PRUNE_FILTER" \
+      --keep-storage="$BUILDX_PRUNE_KEEP_STORAGE"
+  fi
 }
 
 function usage() {
@@ -29,6 +32,9 @@ Description:
     Build Docker image using BuildX
 
 OPTIONS:
+    --builder <string>
+        Override the configured builder instance (default "$BUILDX_BUILDER_NAME")
+
     --platform <string>
         Set target platform for build (format: "linux/amd64,linux/arm64")
 
@@ -37,6 +43,9 @@ OPTIONS:
 
     --push
         Push image to the registry
+
+    --prune
+        Remove build cache
 
     --tag <string>
         Name and optionally a tag
@@ -48,6 +57,10 @@ USAGE
 
 while (("$#")); do
   case "$1" in
+  --builder)
+    BUILDX_BUILDER_NAME="$2"
+    shift 2
+    ;;
   --load | --push)
     BUILDX_ARGS+=" $1"
     shift
@@ -55,6 +68,10 @@ while (("$#")); do
   --platform | --tag)
     BUILDX_ARGS+=" $1=$2"
     shift 2
+    ;;
+  --prune)
+    BUILDX_PRUNE="true"
+    shift
     ;;
   -h | --help)
     usage
@@ -69,16 +86,13 @@ while (("$#")); do
 done
 
 set +e
-docker buildx use $BUILDX_BUILDER_NAME 2>/dev/null
-if [[ $? -ne 0 ]];then
-  set -e
-  docker buildx create \
-    --use \
-    --platform="$BUILDX_BUILDER_PLATFORM" \
-    --name="$BUILDX_BUILDER_NAME"
-fi
+docker buildx create \
+  --platform="$BUILDX_BUILDER_PLATFORM" \
+  --name="$BUILDX_BUILDER_NAME" 2>/dev/null
+set -e
 
 trap _exit EXIT
 
 docker buildx build \
+  --builder=$BUILDX_BUILDER_NAME \
   $BUILDX_ARGS $BUILDX_CONTEXT
